Lesson 5: Part II. Positioning with inline-block
Although the lesson mentioned some layouts with inline-block, the
options given to eliminate white-space were strictly dependent on HTML
actual format.

The conclusion I personally made was to just document it but will ignore
as it is a really flaky use-case that can easily break with code
formatters such as prettier / tidy.